+++ R/gha_update_docs.R
@@ -75,9 +75,12 @@ gha_update_docs <-
 
 
     if (result) {
-      cat(crayon::green("Success\n"))
+      cli::cli_alert_success(paste(
+        cli::col_green("Action updated correctly."),
+        "See {.file {file.path(destdir, basename(filepath))}}"
+      ))
     } else {
-      cat(crayon::magenta("File not updated\n"))
+      cli::cli_alert_danger(cli::col_red("File not updated"))
       return(invisible())
     }
 
@@ -100,11 +103,9 @@ gha_update_docs <-
 
     writeLines(add_platform, con = file.path(destdir, "update-docs.yaml"))
 
-    cat(crayon::green(
-      paste0(
-        "Your package would be deployed on ", platform, "-", version,
-        "\n"
-      )
+    cli::cli_alert_info(paste(
+      "Your package would be deployed on",
+      "{.val {paste0(platform, '-', version)}}"
     ))