+++ NEWS.md
@@ -1,10 +1,12 @@
 # pkgdev (development version)
 
-* Add `pkgdown` to Imports
-* Add `.imgbotconfig` to `.Rbuildignore`.
+-   Add `pkgdown` to Imports
+-   Add `.imgbotconfig` to `.Rbuildignore`.
+-   Remove `crayon` and use `cli` instead.
+-   Crop the pngs of the precomputed vignettes with `knitr::plot_crop()`
 
 # pkgdev 0.1.0
 
 # pkgdev 0.0.1.9000
 
-* Added a `NEWS.md` file to track changes to the package.
+-   Added a `NEWS.md` file to track changes to the package.